[NFC] Address more bit-field storage sizes
Follow on work from #139825.

As previously this is fixing a few more cases where adjacent bit-fields
have types with different base storage sizes, as the MS ABI will not
pack those bit-fields.

+++ lldb/source/Plugins/Language/ObjC/NSArray.cpp
@@ -101,7 +101,7 @@ namespace Foundation1010 {
       uint32_t _used;
       uint32_t _offset;
       uint32_t _size : 28;
-      uint64_t _priv1 : 4;
+      uint32_t _priv1 : 4;
       uint32_t _priv2;
       uint32_t _data;
     };
